add common request
